refactor: adds automation to pull down official translate icon for use in extension
Adds an automation script to cURL the official Amazon Translate icon and add it to the codebase
prior to build. This is done because of legal implications of including the icon with the project
because it is open source. Refactors some of the CSS and moves the Spinner component from the popup
folder to the components folder. Adds SVG import support to vite.

fix #14