feat: adds keyboard shortcut for starting a translation (#32)
* feat: adds option for setting default source language

Adds an option to set the default source language in the extension settings. The popup will default
to the value of the default source language for the source language field.

fix #23

* feat: adds keyboard shortcut for starting a translation

Adds support for translating a web page using the keyboard shortcut Alt+Shift+T and translates using
the configured default source language and default target language. Translations now display an
indicator on the bottom left of the screen while translating the page.

fix #27

* fix: fixes bug with unit test for destroying the overlay